+++ spec/ruby/core/objectspace/weakkeymap/getkey_spec.rb
@@ -9,6 +9,9 @@ ruby_version_is "3.3" do
       map[key1] = true
       map.getkey(key2).should equal(key1)
       map.getkey("X").should == nil
+
+      key1.should == "A" # keep the key alive until here to keep the map entry
+      key2.should == "A" # keep the key alive until here to keep the map entry
     end
 
     it "returns nil when a key cannot be garbage collected" do