+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Compatibility:
 
 * Implement `StringScanner#{peek_byte,scan_byte,scan_integer,named_captures}` methods (#3788, @andrykonchin).
 * Support String patterns in `StringScanner#{exist?,scan_until,skip_until,check_until,search_full}` methods (@andrykonchin).
+* Implement `ObjectSpace::WeakKeyMap` (#3681, @andrykonchin).
 
 Performance: